+++ src/org/gzhmc/report4gzhmc/controller/manage/GradeController.java
@@ -1,6 +1,5 @@
 package org.gzhmc.report4gzhmc.controller.manage;
 
-
 import java.util.List;
 
 import org.gzhmc.common.exception.WebException;
@@ -11,11 +10,13 @@ import javax.servlet.http.HttpServletResponse;
 import org.gzhmc.common.base.BaseController;
 
 import org.gzhmc.report4gzhmc.mapper.CollegeMapper;
+import org.gzhmc.report4gzhmc.mapper.GradeExamMapper;
 import org.gzhmc.report4gzhmc.mapper.GradeMajorCollegeMapper;
 import org.gzhmc.report4gzhmc.mapper.GradeMapper;
 import org.gzhmc.report4gzhmc.mapper.MajorMapper;
 import org.gzhmc.report4gzhmc.model.College;
 import org.gzhmc.report4gzhmc.model.Grade;
+import org.gzhmc.report4gzhmc.model.GradeExam;
 import org.gzhmc.report4gzhmc.model.GradeMajorCollege;
 import org.gzhmc.report4gzhmc.model.Major;
 import org.gzhmc.report4gzhmc.model.ResultJson;
@@ -24,7 +25,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-
 
 /**
  * 班级管理
@@ -39,6 +39,8 @@ public class GradeController extends BaseController {
 	@Autowired
 	GradeMajorCollegeMapper gradeMajorCollegeMapper;
 	@Autowired
+	GradeExamMapper gradeExamMapper;
+	@Autowired
 	MajorMapper majorMapper;
 	@Autowired
 	CollegeMapper collegeMapper;
@@ -50,6 +52,13 @@ public class GradeController extends BaseController {
 	public ModelAndView index(HttpServletRequest request, HttpServletResponse response) {
 		List<GradeMajorCollege> gradeMajorColleges = gradeMajorCollegeMapper.getAll();
 		return new ModelAndView("manage/indexGrade").addObject("gmclist", gradeMajorColleges);
+	}
+
+	// 显示所有实验班级关联信息
+	@RequestMapping(value = { "indexGradeExam", "indexGradeExam.html" })
+	public ModelAndView indexGradeExam(HttpServletRequest request, HttpServletResponse response) {
+		List<GradeExam> gradeExams = gradeExamMapper.getAll();
+		return new ModelAndView("manage/indexGradeExam").addObject("gradeExams", gradeExams);
 	}
 
 	// 跳转到增加或修改信息的页面
@@ -73,7 +82,7 @@ public class GradeController extends BaseController {
 	// 根据id是否存在判断是否为增加或修改信息，进行增加或修改
 	@RequestMapping(value = { "/addGrade.action" })
 	@Transactional
-	public ModelAndView addCollegeAction(HttpServletRequest request, HttpServletResponse response) throws WebException {
+	public void addGradeAction(HttpServletRequest request, HttpServletResponse response) throws WebException {
 		// 获取表单中的值，如果id值为空，则表示新增，否则表示根据此id来修改
 		String id = request.getParameter("id");
 		String cYearClass = request.getParameter("cYearClass");
@@ -86,7 +95,7 @@ public class GradeController extends BaseController {
 		grade.setcCollegeId(StringUtils.string2int(cCollege));
 		grade.setcMajorId(StringUtils.string2int(cMajor));
 		grade.setcYearClass(cYearClass);
-
+		ResultJson json = new ResultJson();
 		int result;
 		if (StringUtils.isNotEmpty(id)) {
 			grade.setcId(StringUtils.string2int(id));
@@ -96,16 +105,18 @@ public class GradeController extends BaseController {
 		}
 		if (result == 1) {
 			// 重定向转到管理班级的页面
-			return new ModelAndView("redirect:/manage/indexGrade.html");
+			json.setSuccess(true);
+			json.setMsg("操作成功!请返回原来页面并刷新。");
 		} else {
 			throw new WebException();
 		}
+		writeResultJson(response, json);
 	}
 
-	// 根据id删除信息
+	// 根据id删除班级信息
 	@RequestMapping(value = { "/delGrade.action" })
 	@Transactional // 增删改操作一定要添加事务回滚
-	public void delMajorAction(HttpServletRequest request, HttpServletResponse response) {
+	public void delGradeAction(HttpServletRequest request, HttpServletResponse response) {
 		// 根据id删除对应的班级
 		String ids = request.getParameter("ids");
 		ResultJson json = new ResultJson();
@@ -113,10 +124,16 @@ public class GradeController extends BaseController {
 		int result;
 		if (idString != null && idString.length > 0) {
 			for (int i = 0; i < idString.length; i++) {
-				result = gradeMapper.delete(StringUtils.string2int(idString[i]));
-				if (result != 1) {
+				if (qualify(StringUtils.string2int(idString[i]))) {
+					result = gradeMapper.delete(StringUtils.string2int(idString[i]));
+					if (result != 1) {
+						json.setSuccess(false);
+						json.setMsg("删除失败!");
+						break;
+					}
+				} else {
 					json.setSuccess(false);
-					json.setMsg("删除失败!");
+					json.setMsg("删除的班级关联其他数据，不允许删除!");
 					break;
 				}
 			}
@@ -125,4 +142,58 @@ public class GradeController extends BaseController {
 		writeResultJson(response, json);
 	}
 
+	// 判断本表关联其他表的数据是否存在，若存在则不允许删除
+	private boolean qualify(int cGradeId) {
+		if (gradeMapper.getByStuGradeId(cGradeId) == 0 && gradeMapper.getByGraExamGradeId(cGradeId) == 0
+				&& gradeMapper.getByteaGradeGradeId(cGradeId) == 0)
+			return true;
+		else
+			return false;
+	}
+
+	//判断是否能够删除班级实验关联信息
+		protected boolean qualifyDelete(int id) {
+		    GradeExam gradeExam=gradeExamMapper.getById(id);
+		    //根据班级实验关联信息查询是否存在报告数据,若存在返回false，即不能删除数据
+			int result=gradeExamMapper.getReportNumByExperimentIdandGradeId(gradeExam.getcGradeId(),gradeExam.getcExperimentId());
+			if(result==0)
+				return true;
+			else
+				return false;
+		}
+
+	/**
+	 * 根据id删除实验班级关联信息
+	 * 
+	 * @param request
+	 * @param response
+	 */
+	@RequestMapping(value = { "/delExmGrade.action" })
+	@Transactional // 增删改操作一定要添加事务回滚
+	public void delExmGreadeAction(HttpServletRequest request, HttpServletResponse response) {
+		// 根据id删除对应的学院
+		String ids = request.getParameter("ids");
+		int id;
+		ResultJson json = new ResultJson();
+		String[] idString = ids.split(",");
+		int result;
+		if (idString != null && idString.length > 0) {
+			for (int i = 0; i < idString.length; i++) {
+				id = StringUtils.string2int(idString[i]);
+				if (qualifyDelete(id)) {
+					result = gradeExamMapper.deleteByPrimaryKey(id);
+					if (result != 1) {
+						json.setSuccess(false);
+						json.setMsg("删除失败!");
+						break;
+					}
+				} else {
+					json.setSuccess(false);
+					json.setMsg("该数据与其他数据关联，删除失败!");
+					break;
+				}
+			}
+		}
+		writeResultJson(response, json);
+	}
 }
